This commit was manufactured by cvs2svn to create tag 'Ext-5-1-2-pre1'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-1-2-pre1@10387 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/Rmon/agutil.c
@@ -193,7 +193,7 @@ AGUTIL_get_oid_value(u_char * var_val, u_char var_val_type,
     var_val_len /= sizeof(oid);
     data_source_ptr->length = var_val_len;
     oid_var = (oid *) var_val;
-    for (iii = 0; iii < data_source_ptr->length; iii++)
+    for (iii = 0; iii < (int)data_source_ptr->length; iii++)
         data_source_ptr->objid[iii] = oid_var[iii];
 
     return SNMP_ERR_NOERROR;
